Allow fractional epochs.

+++ finetune.py
@@ -121,7 +121,7 @@ def train(
         # training hyperparams
         batch_size: int = 128,
         micro_batch_size: int = 4,
-        num_epochs: int = 3,
+        num_epochs: float = 3,
         learning_rate: float = 3e-4,
         cutoff_len: int = 256,
         val_set_size: int = 2000,
